Set security headers in production
There's a list of them here
- https://docs.djangoproject.com/en/3.0/howto/deployment/checklist/#https

And also a command you can run:

`python manage.py check --deploy`

Basically, if we're running on an HTTPS domain, they should all
be set. If not, we don't want to set them as they won't run locally.

I was thinking of making this the same as DJANGO_DEBUG, but we
do want to be able to run this in Prod with DEBUG on (for testing),
and we also want to be able to run this in dev with DEBUG off
(to see if our prod changes will work).

So, another env var.