Create a language-switching link for the top nav
There's a Django form example that does language-switching here:
- https://docs.djangoproject.com/en/3.0/topics/i18n/translation/#the-set-language-redirect-view

And here's the docs on the tags I used:
- https://docs.djangoproject.com/en/3.0/topics/i18n/translation/#other-tags

Add CSS to make it look like a link, so it blends in.

It's actually a button because it sends a post request to the server
to change the language. This is the same pattern we've used in other
apps, such as the CRA app.

Also removed the /translate subdomain because we don't need a
separate page just for translations.