Add "black" code formatter
Black is the prettier of Python, from what I can tell.

It reformats files how it likes, which means some big diffs initially.

Added to pipenv as well so that we can eventually get it into our Github actions.

Apparently the Django team has recently agreed to use it.

Docs are here:
- https://black.readthedocs.io/en/stable/