Move scss back into /static folder but ignore them when building
Previously, I had moved the SCSS files out of the static folder
because they were getting coped over when the app was being built and all of our CSS files were available to be served.

No harm done, really but we shouldn't be serving them.

Then I found the "--watch" option of "django-sass" stopped working
and the docs say it looks in the static folders by default.
(Not sure why when you tell it what folder to look in but there
you go.)

So I moved them back into the static folder, but am telling
"collectstatic" to ignore the "scss" folder when it builds the app,
meaning we won't copy and serve them when we deploy.

Also added `pipenv run css` and `pipenv run csswatch` commands that
we can run to not have to use the CSS path everytime.