Remove |safe and |escape filters on form templates
Was following the default documentation for Django form rendering
and they use the "|escape" filter for error message rendering.

- https://docs.djangoproject.com/en/3.0/topics/forms/#rendering-form-error-messages

However, autoescaping is turned on by default in Django, so it doesn
't
seem like "|escape" is really doing anything.

Tested putting unsafe characters in error messages and they
display the same with or without "|escape", so we don't need it.

The only way to interpret HTML tags and such is to use "|safe".