Added "Invite" view only for admins
Added an in-app form where users can be invited to the app.

Only admins or superusers are able to see this form.

Works very similar to in the admin, except it doesn't let you
select who you are receiving the invitation from. The id of the
currently-logged in user is the one sending the invite.

Went through the bee-keeper code for help here.

Sources:
- https://github.com/bee-keeper/django-invitations/blob/9069002f1a0572ae37ffec21ea72f66345a8276f/invitations/forms.py#L60