Add hidden fields to form but don't display them
Previously, hidden fields would be treated like other fields
and their labels would be visible.

With hidden fields, we want them in the form but we don't want
to have error messages or help text or labels or anything.

Stuck pretty close to the django example code on this one.

Source:
- https://docs.djangoproject.com/en/3.0/topics/forms/#looping-over-hidden-and-visible-fields