Disable button on form submit
The invite view in particular takes a long time (2-4 seconds)
to send an email, and while that's happening, there's no visual
feedback and the button is still active.

When you press the button a second time, the app breaks, so we
don't want people pressing it twice.

Button is un-disabled on every pageload, and form submits always result
in a pageload so I think we're okay here.