Add code instructions page with placeholder images
The markup is a bit weird, but that's because the layout is weird
anyway. The whole page is a big nested list, with sub-instructions
being yet another list.

Added a placekitten image while waiting on the screenshots to
come thru.

Used the same view as for the "number" page, as the variables it
needs are identical.

Sources:
- Nested HTML list: https://stackoverflow.com/questions/5899337/proper-way-to-make-html-nested-list
- https://placekitten.com/g/300/400