Evaluate DJANGO_ADMINS as a list of tuples
As pointed out by @bryan-robitaille, `ADMINS` is meant to be a list
of tuples. For example:

[('John', 'john@example.com'), ('Mary', 'mary@example.com')]

Added code to parse a string formatted to be a tuple.

eg, this will work:

'[("Paul", "paul@example.com"),("Bryan", "bryan@example.com")]'

The `literal_eval` documentation says it's safe, and we control
the environment variable, so we should be okay to ship this as is.
- https://docs.python.org/3/library/ast.html#ast.literal_eval

If the env variable exists but isn't formatted correctly, it will
cause an error on startup, which is what we want, so that we don't
think we've set it and it stops working.