Subject line of invitation emails says what environment it's running in
We are using the "sites" object to put in the domain and name of our
site in most environments (Testing, Staging, Prod), but it's not set
in PR apps (because they disappear), and it's not set locally
(because who cares).

So the `get_site_name` method will return the environment in square brakets
and then "COVID Alert Portal".

If we see "example.com", we look for the Heroku env var and then the
environment becomes the name of the PR app.

This means that if we're in prod or staging without the "Sites" object
set, emails will return Localhost, but this is a pretty minor problem
for us and will be easy to fix if it appears.