Update READMEs
Removed most of the text in the Infrastructure README that talked
about the 3 Covid Shield services working together and some of the
Github Actions that we've removed.

The manual deployment steps are documented in a Google Doc that
people on the team have access to.

Also updated the main README as it was still talking about COVID Shield.