More spacing under the login form
Added 30px of space between the button and "forgot password", and
60px of space between "forgot password" and the little explainer
sentence about healthcare workers.

Conceptually, the form is a single unit, and the "Forgot password"
link is kind of related to the form, so we should group them
close to each other (but not super close).

Whereas the other trailing sentence isn't really related at all,
so let's push that all the way under.