Remove explicit 2FA call after account created
Creating a new account triggers `update` in signals.py, which sends
a 2FA SMS to the new phone number.

We were also explicitly creating an SMS to send at the end of user
creation.

The result was that users would get 2 TFA codes on account creation,
but only the second one was usable.

Removing the account-creation code, leaving the signals one.