Fix main nav layout on login page
On the login page, we zeroed out all the max-width spacing so that
we could have full-bleed columns, but when you are on the login
page _and you are logged in_, the nav menu is touching the side
of the screen, which is a total oversight.

By making our CSS rules more specific, we fix the layout glitch.