Split "push" commands into separate GitHub Actions
Use one action to push to staging and another one to push to prod

Also created a unified action for "build container" so that we're
not building it twice.

Should be good to go here.

The prod one checks for a change to the "VERSION" file before it will
start.