ThrottleMixin uses `dispatch` method, not `render_to_response`
Using mixins allows us to intercept the lifecyle methods of a given
view.

The problem with our ThrottleMixin is that the "render_to_response"
method comes too late in the lifecycle of the class view.

In our COVID key view, the "post" method is called every time before
the "render_to_response" method runs. Therefore, a key is always being
generated, even when the throttling succeeds and tells the user they
can't have any more keys.

By using the "dispatch" method instead, it runs before the "get" or
"post" methods of the View classes, so if the throttling succeeds,
a key is no longer being created.

Sources:
- https://docs.djangoproject.com/en/3.1/ref/class-based-views/base/#django.views.generic.base.View.dispatch