Add new page to generate backup codes when logging in
Added a "No mobile phone button" and a new screen that generates codes.

Techinically this should be a POST request because "GET"ting more
codes destroys the previous codes, so it's not just a read. However,
it was a bit of a conundrum because the other button was already in
a form.

In order to keep the buttons beside each other without resorting to
some wierd positioning, I just created the second button as link.

Nested forms are bad HTML:

```
<form action="/url1">
<button type="submit">Submit</button>
<form action="/url2"><button type="submit>No? Do this other thing</button></form>
</form>
```

There's apparently a form attribute that you can set on input elements, specifiying the form to activate:
- https://www.sitepoint.com/the-html5-form-attribute/

However, it's not supported in IE11 or lower, and we can't realistically
drop support for IE11.

So yeah, made it a GET request -- seemed like the best of bad options.

---

Also changed some of the URL names that were using hyphens so that
they use underscores now.

Changed:
- resend-2fa → resend_2fa
- login-2fa  → login_2fa
- signup-2fa → signup_2fa