Set phone number as an optional field
In anticipation of users signing up but not needing to
enter their phone numbers, we are making "phone number"
an optional field.

Users and admins will not be able to remove phone numbers once
set, but are able to create accounts without initally setting
a phone number.

However, superadmins are able to remove phone numbers using
the Django admin backend.

This change includes a data migration.